src/requirements.txt
anthropic-bedrock==0.8.0
boto3==1.35.88
botocore==1.35.88
FalkorDB==1.0.10
FlagEmbedding==1.3.4
huggingface-hub==0.28.1
json2xml==5.0.4
llama-index-core==0.12.6
llama-index-embeddings-bedrock==0.4.0
llama-index-llms-anthropic==0.6.3
llama-index-llms-bedrock==0.3.3
llama-index-readers-file==0.4.0
llama-index-readers-web==0.3.0
llama-index-vector-stores-opensearch==0.5.0
lru-dict==1.3.0
opensearch-py==2.8.0
pgvector>=0.2.0
pipe==2.2
protobuf==4.25.3
psycopg>=3.1.0
pynvml==11.5.3
pytest==8.2.2
python-dotenv==1.0.1
redis==5.2.1
sentencepiece==0.2.0
sentence-transformers==3.1.1
smart_open==7.0.4
spacy==3.7.5
torch==2.4.1
transformers==4.48.0
tfidf_matcher==0.3.0